refactor: reuse create release logic across clients
